requirements.txt
boto3
faker
pandas
python-dotenv
